% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdcr-package.R
\docType{package}
\name{ggdcr-package}
\alias{ggdcr}
\alias{ggdcr-package}
\title{ggdcr: Downloads data from the Groningen Growth and Development Centre (GGDC)}
\description{
This package intends to simplify the download and data minipulation from the tables avaiable at the GGDC website.
}
\section{\code{ggdcr} Functions}{


\itemize{

 \item \code{\link{maddison_project}}
 \item \code{\link{penn_world_table}}
 \item \code{\link{productivity_level_database}}
 \item \code{\link{ten_sector_database}}

}
}

\keyword{internal}
