% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity_level_database.R
\name{productivity_level_database}
\alias{productivity_level_database}
\title{GGDC Productivity Level Database}
\usage{
productivity_level_database(sheet)
}
\arguments{
\item{sheet}{a number from \code{1} to \code{5}.}
}
\value{
A tidy \code{tibble}
}
\description{
The GGDC Productivity Level Database provides data on relative prices and labor productivity across countries for 42 major economies and up to 35 industries in 2005.
}
\details{
These data complement the World Input-Output Database and are largely based on the results of the 2005 International Comparisons Program (ICP), which estimates purchasing power parities (PPPs) for a global sample of countries.

The description of each worksheet is as follows:

\itemize{

\item \code{1}: GDP and major sector value added relative prices for 2005, USA GDP = 1;
  \item \code{2}: GDP and major sector labor productivity (value added per hour worked) for 2005, USA = 1;
  \item \code{3}: GDP and major sector gross output relative prices for 2005, USA GDP = 1;
  \item \code{4}: Sector gross output relative prices for 2005, USA GDP = 1; and
  \item \code{5}: Industry gross output relative prices for 2005, USA GDP = 1.

}
}
\examples{
library(ggdcr)
productivity_level_database(2)
}
\references{
Robert Inklaar and Marcel P. Timmer (2014), \emph{"The Relative Price of Services"}, Review of Income and Wealth 60(4): 727-746, avaiable at: \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/roiw.12012}.
}
