% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socio_economic_accounts.R
\name{socio_economic_accounts}
\alias{socio_economic_accounts}
\title{World Input-Output Database}
\usage{
socio_economic_accounts()
}
\value{
A tidy \code{tibble}.
}
\description{
The World Input-Output Database (WIOD) consists of a series of databases on output, prices, capital stocks, and employment. It covers 28 EU countries and 15 other major countries in the world for the period from 2000 to 2014.
}
\examples{
library(ggdcr)
socio_economic_accounts()
}
\references{
Timmer, M. P., Dietzenbacher, E., Los, B., Stehrer, R. and de Vries, G. J. (2015), \emph{"An Illustrated User Guide to the World Input-Output Database: The Case of Global Automotive Production"}, Review of International Economics., 23: 575-605, available at \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/roie.12178}.
}
