% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ten_sector_database.R
\name{ten_sector_database}
\alias{ten_sector_database}
\title{GGDC 10-Sector Database}
\usage{
ten_sector_database()
}
\value{
A "tidy" \code{tibble}.
}
\description{
The GGDC 10-Sector Database provides a long-run internationally comparable dataset on sectoral productivity performance in Africa, Asia, and Latin America. Variables covered in the data set are annual series of value added, output deflators, and persons employed for 10 broad sectors.
}
\details{
The GGDC 10-Sector Database gives sector detail to the historical macro data in Maddison (2003) from 1950 onwards. It consists of series for 11 countries in Africa, 11 countries in Asia, 2 countries in the Middle East and North Africa, and 9 in Latin-America.

It should be stressed that the estimates for the total economy are aggregated across sectors and that, because of adjustments at the sector level, the aggregate results are not fully consistent with the national accounts aggregates (see the sources and methods document). Also note that value added data in this database are expressed in local currencies.

The sectors covered are in the \code{key} column and include:

\itemize{

\item \code{Agriculture}: Agriculture, Hunting and Forestry, Fishing;
  \item \code{Mining}: Mining and Quarrying;
  \item \code{Manufacturing}: Manufacturing;
  \item \code{Public Utilities}: Electricity, Gas and Water supply;
  \item \code{Construction}: Construction;
  \item \code{Retail}: Wholesale and Retail trade; repair of motor vehicles, motorcycles and personal and household goods, Hotels and Restaurants;
  \item \code{Transport}: Transport, Storage and Communications;
  \item \code{Business}: Financial Intermediation, Renting and Business Activities (excluding owner occupied rents);
  \item \code{Government}: Public Administration and Defense, Education, Health and Social work; and
  \item \code{Other}: Other Community, Social and Personal service activities, Activities of Private Households.

}
}
\examples{
library(ggdcr)
ten_sector_database()
}
\references{
\itemize{

  \item Timmer, M.P., de Vries, G.J., & de Vries, K. (2015). \emph{Patterns of Structural Change in Developing Countries.} . In J. Weiss, & M. Tribe (Eds.), Routledge Handbook of Industry and Development. (pp. 65-83). Routledge.

  \item The working paper can be found at \url{www.ggdc.net/publications/memorandum/gd149.pdf}

  \item Sources and methods used for calculations: \url{https://www.rug.nl/ggdc/docs/10sector_sm_jan2015.pdf}

}
}
