% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivity_level_database.R
\name{productivity_level_database}
\alias{productivity_level_database}
\title{GGDC Productivity Level Database}
\usage{
productivity_level_database(sheet)
}
\arguments{
\item{sheet}{a number ranging from \code{1} to \code{5}.}
}
\value{
A tidy \code{tibble}
}
\description{
The GGDC Productivity Level Database provides data on relative prices and labor productivity across countries for 42 major economies and up to 35 industries in 2005.
}
\details{
These data complement the World Input-Output Database and are largely based on the results of the 2005 International Comparisons Program (ICP), which estimates purchasing power parities (PPPs) for a global sample of countries.

The description of each worksheet is as follows:

\itemize{

\item \code{1}: GDP and major sector value added relative prices for 2005, USA GDP = 1;
  \item \code{2}: GDP and major sector labor productivity (value added per hour worked) for 2005, USA = 1;
  \item \code{3}: GDP and major sector gross output relative prices for 2005, USA GDP = 1;
  \item \code{4}: Sector gross output relative prices for 2005, USA GDP = 1; and
  \item \code{5}: Industry gross output relative prices for 2005, USA GDP = 1.

}
}
\note{
Except in sheet \code{2}, all data are relative prices, so the purchasing power parity (PPP) divided by the exchange rate. All relative prices are relative to the USA GDP price level, shown in sheet \code{1}. Labor productivity is defined as PPP-converted value added divided by the number of hours worked. PPPs used are those shown in sheet \code{1}.

Number of hours worked is from the WIOD Socio-Economic Accounts, except for Argentina, Chile, and South Africa, where data on the number of persons engaged is from the 10-Sector Database and the Africa Sector Database. The average number of hours worked in South Africa is assumed to be equal to that in the USA in every sector. The average number of hours worked in Argentina and Chile is taken from PWT8.0 and assumed to be equal across sectors.
}
\examples{
library(ggdcr)
productivity_level_database(2)
}
\references{
Robert Inklaar and Marcel P. Timmer (2014), \emph{"The Relative Price of Services"}, Review of Income and Wealth 60(4): 727-746, available at: \url{https://onlinelibrary.wiley.com/doi/abs/10.1111/roiw.12012}.
}
